# Improve hashing speed