--- a/hash/hash_test.go
+++ b/hash/hash_test.go
@@ -37,3 +37,31 @@ func TestGenMerkleHashFunc(t *testing.T) {
 	// different children (e.g. different order) -> different hash
 	r.NotEqual(GenMerkleHashFunc(aChallenge)(nil, lChild, rChild), GenMerkleHashFunc(aChallenge)(nil, rChild, lChild))
 }
+
+func TestGenLabelHashFuncHash(t *testing.T) {
+	challenge := []byte("a123123123mskofvuw098e71bnj91273")
+	data := []byte("12312390819023879012379184718920408data0123128991231239081")
+
+	require.Equal(
+		t,
+		[]byte{
+			104, 1, 230, 180, 3, 202, 139, 10, 164, 69, 74, 169, 116, 77, 27, 134,
+			138, 163, 204, 32, 182, 197, 205, 194, 13, 48, 30, 224, 254, 1, 46, 133,
+		},
+		GenLabelHashFunc(challenge)(data),
+	)
+}
+
+func TestGenMekleHashFuncHash(t *testing.T) {
+	challenge := []byte("a123123123mskofvuw098e71bnj91273")
+	lChild, rChild := []byte("left-one"), []byte("right-one")
+
+	require.Equal(
+		t,
+		[]byte{
+			103, 1, 150, 181, 115, 169, 147, 120, 81, 61, 244, 238, 253, 189, 132, 180,
+			181, 155, 232, 100, 130, 168, 223, 56, 2, 186, 250, 51, 76, 118, 6, 41,
+		},
+		GenMerkleHashFunc(challenge)(nil, lChild, rChild),
+	)
+}