--- a/hash/hash_test.go
+++ b/hash/hash_test.go
@@ -3,10 +3,12 @@ package hash
 import (
 	"testing"
 
+	"github.com/spacemeshos/sha256-simd"
 	"github.com/stretchr/testify/require"
 )
 
 func TestGenLabelHashFunc(t *testing.T) {
+	t.Parallel()
 	r := require.New(t)
 
 	aChallenge, bChallenge := []byte("a"), []byte("b")
@@ -23,6 +25,7 @@ func TestGenLabelHashFunc(t *testing.T) {
 }
 
 func TestGenMerkleHashFunc(t *testing.T) {
+	t.Parallel()
 	r := require.New(t)
 
 	aChallenge, bChallenge := []byte("a"), []byte("b")
@@ -39,29 +42,75 @@ func TestGenMerkleHashFunc(t *testing.T) {
 }
 
 func TestGenLabelHashFuncHash(t *testing.T) {
+	t.Parallel()
+	req := require.New(t)
 	challenge := []byte("a123123123mskofvuw098e71bnj91273")
 	data := []byte("12312390819023879012379184718920408data0123128991231239081")
 
-	require.Equal(
-		t,
-		[]byte{
-			104, 1, 230, 180, 3, 202, 139, 10, 164, 69, 74, 169, 116, 77, 27, 134,
-			138, 163, 204, 32, 182, 197, 205, 194, 13, 48, 30, 224, 254, 1, 46, 133,
-		},
-		GenLabelHashFunc(challenge)(data),
-	)
+	expected := []byte{
+		104, 1, 230, 180, 3, 202, 139, 10, 164, 69, 74, 169, 116, 77, 27, 134,
+		138, 163, 204, 32, 182, 197, 205, 194, 13, 48, 30, 224, 254, 1, 46, 133,
+	}
+
+	t.Run("generic implementation", func(t *testing.T) {
+		t.Parallel()
+		for i := 0; i < 10; i++ {
+			req.Equal(expected, genLabelHashFuncGeneric(challenge)(data))
+		}
+	})
+	t.Run("minio implementation", func(t *testing.T) {
+		t.Parallel()
+		if _, ok := sha256.New().(*sha256.Digest); !ok {
+			t.Skip("SHA extensions are not supported")
+		}
+		for i := 0; i < 10; i++ {
+			req.Equal(expected, genLabelHashFuncMinio(challenge)(data))
+		}
+	})
+	t.Run("public hash func generator", func(t *testing.T) {
+		t.Parallel()
+		for i := 0; i < 10; i++ {
+			req.Equal(expected, GenLabelHashFunc(challenge)(data))
+		}
+	})
 }
 
 func TestGenMekleHashFuncHash(t *testing.T) {
+	t.Parallel()
+	req := require.New(t)
 	challenge := []byte("a123123123mskofvuw098e71bnj91273")
 	lChild, rChild := []byte("left-one"), []byte("right-one")
 
-	require.Equal(
-		t,
-		[]byte{
-			103, 1, 150, 181, 115, 169, 147, 120, 81, 61, 244, 238, 253, 189, 132, 180,
-			181, 155, 232, 100, 130, 168, 223, 56, 2, 186, 250, 51, 76, 118, 6, 41,
-		},
-		GenMerkleHashFunc(challenge)(nil, lChild, rChild),
-	)
+	expected := []byte{
+		103, 1, 150, 181, 115, 169, 147, 120, 81, 61, 244, 238, 253, 189, 132, 180,
+		181, 155, 232, 100, 130, 168, 223, 56, 2, 186, 250, 51, 76, 118, 6, 41,
+	}
+
+	t.Run("generic implementation", func(t *testing.T) {
+		t.Parallel()
+		var digest []byte
+		for i := 0; i < 10; i++ {
+			digest = genMerkleHashFuncGeneric(challenge)(digest, lChild, rChild)
+			req.Equal(expected, digest)
+		}
+	})
+	t.Run("minio implementation", func(t *testing.T) {
+		t.Parallel()
+		if _, ok := sha256.New().(*sha256.Digest); !ok {
+			t.Skip("SHA extensions are not supported")
+		}
+		var digest []byte
+		for i := 0; i < 10; i++ {
+			digest = genMerkleHashFuncMinio(challenge)(digest, lChild, rChild)
+			req.Equal(expected, digest)
+		}
+	})
+	t.Run("public hash func generator", func(t *testing.T) {
+		t.Parallel()
+		var digest []byte
+		for i := 0; i < 10; i++ {
+			digest = GenMerkleHashFunc(challenge)(digest, lChild, rChild)
+			req.Equal(expected, digest)
+		}
+	})
 }