--- a/shared/shared.go
+++ b/shared/shared.go
@@ -4,8 +4,8 @@ import (
 	"encoding/binary"
 	"os"
 
-	"github.com/minio/sha256-simd"
 	"github.com/spacemeshos/merkle-tree"
+	"github.com/spacemeshos/sha256-simd"
 )
 
 //go:generate scalegen -types MerkleProof,ProofMessage,Proof
@@ -18,6 +18,8 @@ const (
 	// OwnerReadWrite is a standard owner read / write file permission.
 	OwnerReadWrite = os.FileMode(0o600)
 )
+
+type LabelHash func(data []byte) []byte
 
 // FiatShamir generates a set of indices to include in a non-interactive proof.
 func FiatShamir(challenge []byte, spaceSize uint64, securityParam uint8) map[uint64]bool {
@@ -44,9 +46,9 @@ func FiatShamir(challenge []byte, spaceSize uint64, securityParam uint8) map[uin
 //
 // ⚠️ The resulting function is NOT thread-safe, however different generated instances are independent.
 // The code is optimized for performance and memory allocations.
-func MakeLabelFunc() func(hash func(data []byte) []byte, labelID uint64, leftSiblings [][]byte) []byte {
+func MakeLabelFunc() func(hash LabelHash, labelID uint64, leftSiblings [][]byte) []byte {
 	var buffer []byte
-	return func(hash func(data []byte) []byte, labelID uint64, leftSiblings [][]byte) []byte {
+	return func(hash LabelHash, labelID uint64, leftSiblings [][]byte) []byte {
 		// Calculate the buffer required size.
 		// 8 is for the size of labelID.
 		// leftSiblings slice might contain nil values, so the result size is inflated and used as an upper bound.