## Changes
- bumped sha256-simd to our fork
- rewritten the tree hashing functions to use our fork
- refactored FiatShamir to better use sha256 to improve speed.

## Tree generation benchmark
### Before:
>BenchmarkGetProof-20    1   30012575868 ns/op   5.53 MB/s   5183499 leafs/op    172783 leafs/sec    977512712 B/op  6295 allocs/op

### After:
>BenchmarkGetProof-20    1   30026978460 ns/op   6.60 MB/s   6193813 leafs/op    206460 leafs/sec    1127514808 B/op 6207 allocs/op

## FiatShamir benchmark
### Before:
>BenchmarkFiatShamir-20   77942   15358 ns/op   8002 B/op   155 allocs/op
### After:
>BenchmarkFiatShamir-20   88592   13230 ns/op   3366 B/op     9 allocs/op